Added comma shorthand for lists/tuples
The comma (,) is now just an alias for cons; wherever it appears in the
program, it is interpreted as "cons". Note that the comma does not need to be
whitespace-delimited like "cons" does, allowing for concise expressions of
lists, such as:

    ,1,2,3,4 empty

Will possibly add similar shorthands for other commonly used functions (a nice
advantage to using almost no punctuation in the language). Perhaps I should
rename the language to golflang.