Added macros, changed comments, added [] lambdas
I added [] lambdas. They work in exactly the same way as () lambdas, except
ints are interpeted as args, and args as ints. Thus, (+ .1 5) and [+ 1 .5] are
equivalent functions; this makes some functions much more concise.

I added the following macros:

    ?       ->      if
    ;       ->      empty
    <       ->      car
    >       ->      cdr

Since ; was the comment character, I simply changed the comment character to be
instead. Now, lists have a very nice and short syntax:

    ,1,2,3,4;

In the future, I may want to change < and > so they can be used to test
equality, but I may just use gt and lt for this anyway