Added string and comparison functions
substr and strcon were added to Init.hs, for getting substrings and
concatenating strings. Moreover, car and cdr were overloaded to be able to
handle strings themselves; car returns a length-1 string representing the first
character, and cdr returns the rest of the string. This has also resulted in
better error messages for car/cdr.

eq, gt, and lt are now all defined for dealing with two Intvals or two Strvals.

The file at examples/findsubstr.fl is an example of how to deal with strings.